LayeredFS: Remove accidental root-entry storage of upper-entries (#62)
An invariant maintained in layered FS is that root entries are ever only
lower/tombstone values. The insertion here is old cruft from before that
invariant was established, and one of the checks for the invariant got
triggered as I was developing another feature. This particular PR fixes
things up by removing the insertion and making the invariant be more
explicit.