Add metadata operations for all file systems (#76)
This PR introduces metadata storage to the file system, allowing storing shim-polymorphic data at either the file OR fd layers, depending on the aliasing needs of the shim.

The main reason to add this came from recognizing that there are situations where we want to add data to the "file struct" (shared across duplicated fds, but not across non-duplicated fds) and to the file itself (shared across all fds on the file), as was noticed in #73. This PR sets up the necessary scaffolding for it.

The way it is shim-polymorphic is by not dictating what types are valid metadata. Instead _any_ type is valid metadata to be attached (although only one valid value can be stored for any particular type: the type itself is the key to the metadata maps). Some things may be supported by some file systems, and others might not. For now, all our file systems will support all possible metadata we want to throw into it.

Minor caveat: layered FS has slight oddities in how we'd need to update metadata if the lower layer is read-only. Currently, it simply throws an `unimplemented` exception and would be handled in a separate PR in the future (I am considering this low priority, unless we actually hit such a use case).