Implement fallible memcpy for Linux Kernel platform (#276)
This PR implements `RawConstPointer` and `RawMutPointer` for
`LinuxKernel`. Also fixed an issue where we use `mutate_subslice_with` on user space
pointer, which assumes the underlying memory is present (i.e., no page faults).

Note that read from/write to user space pointer in kernel mode should
not panic. Instead, it should be able to resume the execution and return
an error. Linux achieves this through building an exception table:
https://elixir.bootlin.com/linux/v5.19.17/source/arch/x86/mm/fault.c#L1327,
which we would implement in the next PR.