Update nightly versions for SNP and LVBS (#399)
As discovered in #398, the nightly version we are using on SNP and LVBS
has been lagging behind the version of stable Rust we are targeting.
This PR fixes this up to by updating our version of nightly.

The changes are fairly minimal:
- the target JSON schema has minor updates
- some pedantic clippy warnings have changed
- "similar names" has gotten weaker; so I've changed `expect` to `allow`
(otherwise it will complain about "unfulfilled lint" on stable; I've
added comments to this end)
- the transmute warnings have gotten stronger, so I just fixed up places
it complained by using actually safer versions that check the alignment
and reduce the chances of UB. Indeed, this is partly why I had added a
transmute ratchet in the past (since transmute can often be replaced by
less dangerous operations), and this reduces that transmute ratchet by a
few, which is great! :)

Along the way, this PR also does a couple of minor drive-by changes:
- update the bacon config to more easily allow running checks on SNP (we
already had something similar for LVBS and such, this just rounds it
out)
- the CI definition for SNP had a minor issue (not problematic since the
two files were identical, but nonetheless, good to fix up and make
consistent)