Eliminate StdioFile; manage via LiteBox raw FDs (#398)
This PR sets up the basis for directly using LiteBox raw FDs within the
Linux shim. While other components will be migrated in future PRs, as a
way to demonstrate what this usage looks like, and to validate the
migration, this PR also at the same time eliminates `StdioFile` from the
Linux shim, switching all stdio behavior to using the raw FDs.

Given that this PR is setting up the basis for future PRs that will
migrate components, this one is a little bit on the larger side in terms
of code addition (since it kind of needs to place a bunch of `todo!`s
and such that will be cleared out soon). For future PRs that are
cleaning up the shim, I expect some of the redundant code to go away.
For now, yes, there are some redundancies with (for example) regular
files.

This PR depends on (and is blocked on #393).

---

This is yet another PR in the series of PRs towards the better FD design
 (connected to #31).