Add support for std::spawn and separate stacks in Linux userland (#404)
This PR adds the support for std::spawn and separate stacks in Linux
userland.

The key trick in enabling this support is to stitching the stack
together between `thread_start_asm` and `syscall_callback`.
Specifically, the stack/frame pointers we store in `thread_start_asm`
right before calling `thread_start_internal` is the same as what's
restored in `syscall_callback` right before calling `syscall_handler`.
Furthermore, the stack layout of `thread_start_asm` is the same as
`syscall_callback`, which allows the return to the caller of
`thread_start_asm` when `sys_exit_group` or `sys_exit` is called on the
main/child thread.

I had to disable seccomp tests because it cannot differentiate syscalls
made by the platform's std from those made by the guest.

---------

Co-authored-by: Weidong Cui <Weidong.Cui@microsoft.com>