Separated raw descriptor store + better `close` (#416)
This PR does a bunch of changes:

1. Separated raw descriptor store: this allows us to (eventually) have
task/process-specific FD storage. This does require the shim to store
the raw->typed descriptor table, which technically (as it currently
stands) leads to an additional descriptor table at the shim, but the
plan is that the old descriptor table will go away.
2. Removes the usage of `Weak` and instead just gives strong `Arc`s.
This is more of a drastic change, since it now has the potential
downside of something within LiteBox holding on to an `Arc` for "too
long" and leading (essentially) to a resource leak. However, I suspect
we'll be able to set up some watchers if we notice something like this
happen, and clean that up if needed. However, this simplifies some
reasoning regarding what happens around closed raw FDs (see next point).
3. Closing of raw FDs (i.e., `fd_consume_raw_integer`) no longer can
return a `CurrentlyUnconsumable` but instead always returns an `Arc`.
This means that we no longer need the fueled-loop and cannot get blocked
by other operations. This does need some changes to LiteBox itself (see
next point) but basically, this allows us to emulate Linux semantics
better.
4. Updates the `close` operations in FS/Net in LiteBox to not be
FD-consuming but instead to take a ref. This does have the downside of
us no longer being able to use the FD as a liveness-token, which means
that all operations can now return a "operating on closed FD" error, but
has the upside of allowing for an unblocked `close` operation. For the
FS, this is basically largely straightforward, since the close operation
doesn't _really_ need to do much cleanup. However, for the Net
subsystem, we need more work (see next point).
5. Updates the Net subsystem to support a queued-close, to account for
ongoing operations. There are fast-paths to close out early if we can,
but if something cannot be closed and cleaned up immediately, then it
gets queued into a system that attempts a close on every
network-platform interaction (again with fast-paths to prevent much
slowdown) but this effectively gives us the GC-like behavior we'd like.

---

To review the PR, it is likely easiest to read each commit on its own.

---

This is yet another PR in the series of PRs towards the better FD design
(connected to https://github.com/microsoft/litebox/issues/31).