Support graceful close & fix connect (#497)
When close a socket, we shouldn't remove it from the `socket set`
directly. Instead, we need to keep it around for a while so that it can
finish the `FIN` handshake. The behavior of socket close can be
configured via `setsockopt` with `SO_LINGER`:

```
Enable(l_onoff!=0):
    Timeout (l_linger > 0): The close() call will block until the socket's unsent data is transmitted and acknowledged, or until the timeout is reached. If the timeout expires, the connection is closed forcefully.
    Timeout (l_linger = 0): The close() call will immediately send a TCP RST (reset) packet, terminating the connection abruptly without waiting for data to be sent. This can cause data loss.
Disable (l_onoff=0): The socket is closed gracefully using the standard TCP handshake (FIN sequence) without any specific timeout delay beyond the normal TCP process. 
```

To the end, `close` now accepts additional argument `CloseBehavior`, and
when a socket is closed, we save it to another queue instead of removing
it from `socket_set`.

Following the same pattern, `connect` now can be called multiple times
with `check_progress = true` to check whether a TCP connection is
established, instead of relying on events on a socket.

Fix closing a server socket in `close_handle` where we need to remove
its all listening sockets as well.

This PR also fixes an event issue for socket. A closed socket should
have `Events::HUP`, which is useful when a connection fails (e.g., no
server is listening or receiving invalid packets) and needs to notify
any waiters.